Set mariadb's pid-dir to mysql's datadir
The rationale behind the change is /run owned by root:root/0755 on el7
so the mysql user can't create /var/run/mysql (/var/run is an symlink to
/run)
when mariadb is restarted (or started for the first time after a reboot)
leaving
mariadb in an failed state.

The change should be merged after
https://github.com/nexcess/ansible-role-mariadb/pull/10
so /var/lib/mysql is created by the package like normal.